refactor(node): Move `Console` definition to `console.d.ts`
